Make running queries in multiple databases easier to do.
Store the values needed to run the query in Query Results' buffer-level
variables. Now there is no need to go back to the original SQL file.

The Query Results buffer now appears in a new window instead of
replacing the SQL file in its window. If the buffer is already in a
visible window, no new one is created.