Remove the color from tabline's "I'm Modified" indicator.
It was a pain picking good colors to go with the colorscheme's dark and
light modes. And if I ever change colorschemes, I'll have to do it all
over again. I added another space before the star, and it's just as easy
to see now.

And my setup is 6 lines simpler now. Yay!