sql.vim: Switch to a time limit for the alignment phase.
A very narrow result set will align just as quickly as a moderately
sized one or a wide one with very few rows. Instead of placing a limit
on the number of rows, this change estimates the amount of time to align
the table. A threshold (defaulting to 5 seconds) determines whether to
perform the alignment or not.