Use the auto-detected value of background to match the terminal.
As explained in this neovim issue,
https://github.com/neovim/neovim/issues/14058#issuecomment-793050809,
the timer_start call to run DeferredColorSchemeSet() is a bit of a hack,
but it puts the setting of the colorscheme at the right time to enable
using the auto-detected value of background. Prior to this, I was always
seeing a dark background in neovim, even when the terminal's background
was light.