sql.vim: Change the way database connections are displayed and selected.
The change is to a 2D picker. j/k select the server, and h/l choose the
database on that server to use. There is a new way to add servers and
databases to the list now: 'New...' appears at the end of the row or
column, and selecting that prompts for the new value(s).