Add mappings for horizontal scrolling - <C-s> and <C-d>.
IMHO, (neo)vim is lacking in its implementation of horizontal scrolling.
There is `<count>zl` and `<count>zh`, but that's a lot of typing,
especially when `<C-f>` and `<C-b>` exist for the vertical direction.
This change usurps the `<C-d>` mapping for my own purpose, but I never
really used that or `<C-u>` anyway.