Reinstall vim-rest-console.
I once again need to test REST calls. Why not use Neovim, of course? The
rest-nvim plugin was in a transition toward a more complicated setup
that wasn't working for me. The original author of vim-rest-console
hadn't been maintaining the plugin, so I browsed its forks, and found
this one that's active and has quite a few pull requests merged into it.